requirements.txt
asgiref==3.3.1
blis==0.4.1
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.5
Django==3.1.7
djangorestframework==3.12.2
docx2txt==0.8
drf-yasg==1.20.0
en-core-web-sm==2.3.1
idna==2.10
importlib-metadata==3.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.1
lxml==4.6.2
MarkupSafe==1.1.1
murmurhash==1.0.5
nltk==3.5
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pdfminer.six==20200517
pdfplumber==0.5.27
phonenumbers==8.12.19
Pillow==8.1.1
plac==1.1.3
preshed==3.0.5
pycryptodome==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
regex==2020.11.13
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sortedcontainers==2.3.0
spacy==2.3.1
sqlparse==0.4.1
srsly==1.0.5
stemming==1.0.1
thinc==7.4.1
tika==1.24
tqdm==4.58.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
Wand==0.6.6
wasabi==0.8.2
zipp==3.4.0
